requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
astroid==2.0.3
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jwcrypto==0.5.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pdfkit==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
sightengine==1.4.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.10
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wkhtmltopdf==0.2
wrapt==1.10.11
WTForms==2.2.1
